# erp-domain/openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:51904
  description: Generated server url
paths:
...
  /api/v1/sales/targets/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateTarget
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesTargetRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesTargetDto"
    delete:
      tags:
      - sales-controller
      operationId: deleteTarget
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/sales/promotions/{id}:
    put:
      tags:
      - sales-controller
      operationId: updatePromotion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePromotionDto"
    delete:
      tags:
      - sales-controller
      operationId: deletePromotion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/sales/orders/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesOrderDto"
    delete:
      tags:
      - sales-controller
      operationId: deleteOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/sales/credit-requests/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateCreditRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditRequestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCreditRequestDto"
...
  /api/v1/sales/targets:
    get:
      tags:
      - sales-controller
      operationId: targets
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListSalesTargetDto"
    post:
      tags:
      - sales-controller
      operationId: createTarget
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesTargetRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesTargetDto"
  /api/v1/sales/promotions:
    get:
      tags:
      - sales-controller
      operationId: promotions
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPromotionDto"
    post:
      tags:
      - sales-controller
      operationId: createPromotion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePromotionDto"
  /api/v1/sales/orders:
    get:
      tags:
      - sales-controller
      operationId: orders
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListSalesOrderDto"
    post:
      tags:
      - sales-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesOrderDto"
  /api/v1/sales/orders/{id}/confirm:
    post:
      tags:
      - sales-controller
      operationId: confirmOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesOrderDto"
  /api/v1/sales/orders/{id}/cancel:
    post:
      tags:
      - sales-controller
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesOrderDto"
  /api/v1/sales/dispatch/confirm:
    post:
      tags:
      - sales-controller
      operationId: confirmDispatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchConfirmRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDispatchConfirmResponse"
  /api/v1/sales/credit-requests:
    get:
      tags:
      - sales-controller
      operationId: creditRequests
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListCreditRequestDto"
    post:
      tags:
      - sales-controller
      operationId: createCreditRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditRequestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCreditRequestDto"
...
  /api/v1/orchestrator/factory/dispatch/{batchId}:
    post:
      tags:
      - orchestrator-controller
      operationId: dispatch
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
  /api/v1/dispatch/confirm:
    post:
      tags:
      - dispatch-controller
      operationId: confirmDispatch_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchConfirmationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDispatchConfirmationResponse"
...
  /api/v1/sales/orders/{id}/status:
    patch:
      tags:
      - sales-controller
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSalesOrderDto"
...
  /api/v1/sales/dealers/search:
    get:
      tags:
      - sales-controller
      operationId: searchDealers
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListDealerLookupResponse"
...
  /api/v1/dealers/{dealerId}/ledger:
    get:
      tags:
      - dealer-controller
      operationId: dealerLedger
      parameters:
      - name: dealerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/dealers/{dealerId}/invoices:
    get:
      tags:
      - dealer-controller
      operationId: dealerInvoices_1
      parameters:
      - name: dealerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/dealers/{dealerId}/aging:
    get:
      tags:
      - dealer-controller
      operationId: dealerAging
      parameters:
      - name: dealerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/dealers/search:
    get:
      tags:
      - dealer-controller
      operationId: searchDealers_1
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListDealerLookupResponse"
...
components:
  schemas:
...
    SalesTargetRequest:
      required:
      - name
      - periodEnd
      - periodStart
      - targetAmount
      type: object
      properties:
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        targetAmount:
          type: number
        achievedAmount:
          type: number
        assignee:
          type: string
    ApiResponseSalesTargetDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SalesTargetDto"
        timestamp:
          type: string
          format: date-time
    SalesTargetDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        targetAmount:
          type: number
        achievedAmount:
          type: number
        assignee:
          type: string
    PromotionRequest:
      required:
      - discountType
      - discountValue
      - endDate
      - name
      - startDate
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
    ApiResponsePromotionDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/PromotionDto"
        timestamp:
          type: string
          format: date-time
    PromotionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
    SalesOrderItemRequest:
      required:
      - productCode
      - quantity
      - unitPrice
      type: object
      properties:
        productCode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        gstRate:
          type: number
    SalesOrderRequest:
      required:
      - items
      - totalAmount
      type: object
      properties:
        dealerId:
          type: integer
          format: int64
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderItemRequest"
        gstTreatment:
          type: string
        gstRate:
          type: number
        gstInclusive:
          type: boolean
        idempotencyKey:
          type: string
    ApiResponseSalesOrderDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SalesOrderDto"
        timestamp:
          type: string
          format: date-time
    SalesOrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        orderNumber:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        subtotalAmount:
          type: number
        gstTotal:
          type: number
        gstRate:
          type: number
        gstTreatment:
          type: string
        gstRoundingAdjustment:
          type: number
        currency:
          type: string
        dealerName:
          type: string
        traceId:
          type: string
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderItemDto"
...
    CreditRequestRequest:
      required:
      - amountRequested
      type: object
      properties:
        dealerId:
          type: integer
          format: int64
        amountRequested:
          type: number
        reason:
          type: string
        status:
          type: string
    ApiResponseCreditRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/CreditRequestDto"
        timestamp:
          type: string
          format: date-time
    CreditRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        dealerName:
          type: string
        amountRequested:
          type: number
        status:
          type: string
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CancelRequest:
      type: object
      properties:
        reason:
          type: string
    DispatchConfirmRequest:
      type: object
      properties:
        packingSlipId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DispatchLine"
        adminOverrideCreditLimit:
          type: boolean
    DispatchLine:
      required:
      - shipQty
      type: object
      properties:
        lineId:
          type: integer
          format: int64
        batchId:
          type: integer
          format: int64
        shipQty:
          type: number
        priceOverride:
          type: number
        discount:
          type: number
        taxRate:
          type: number
        taxInclusive:
          type: boolean
...
    ApiResponseDispatchConfirmResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/DispatchConfirmResponse"
        timestamp:
          type: string
          format: date-time
    CogsPostingDto:
      type: object
      properties:
        inventoryAccountId:
          type: integer
          format: int64
        cogsAccountId:
          type: integer
          format: int64
        cost:
          type: number
    DispatchConfirmResponse:
      type: object
      properties:
        packingSlipId:
          type: integer
          format: int64
        salesOrderId:
          type: integer
          format: int64
        finalInvoiceId:
          type: integer
          format: int64
        arJournalEntryId:
          type: integer
          format: int64
        cogsPostings:
          type: array
          items:
            $ref: "#/components/schemas/CogsPostingDto"
        dispatched:
          type: boolean
        arPostings:
          type: array
          items:
            $ref: "#/components/schemas/AccountPostingDto"
...
    DispatchRequest:
      required:
      - batchId
      - postingAmount
      - requestedBy
      type: object
      properties:
        batchId:
          type: string
        requestedBy:
          type: string
        postingAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
...
    DispatchConfirmationRequest:
      required:
      - lines
      - packagingSlipId
      type: object
      properties:
        packagingSlipId:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: "#/components/schemas/LineConfirmation"
        notes:
          type: string
        confirmedBy:
          type: string
    LineConfirmation:
      required:
      - lineId
      - shippedQuantity
      type: object
      properties:
        lineId:
          type: integer
          format: int64
        shippedQuantity:
          type: number
        notes:
          type: string
    ApiResponseDispatchConfirmationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/DispatchConfirmationResponse"
        timestamp:
          type: string
          format: date-time
    DispatchConfirmationResponse:
      type: object
      properties:
        packagingSlipId:
          type: integer
          format: int64
        slipNumber:
          type: string
        status:
          type: string
        confirmedAt:
          type: string
          format: date-time
        confirmedBy:
          type: string
        totalOrderedAmount:
          type: number
        totalShippedAmount:
          type: number
        totalBackorderAmount:
          type: number
        journalEntryId:
          type: integer
          format: int64
        cogsJournalEntryId:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: "#/components/schemas/LineResult"
        backorderSlipId:
          type: integer
          format: int64
...
    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
        timestamp:
          type: string
          format: date-time
...
    StatusRequest:
      type: object
      properties:
        status:
          type: string
...
    ApiResponseListSalesTargetDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesTargetDto"
        timestamp:
          type: string
          format: date-time
    ApiResponseListPromotionDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PromotionDto"
        timestamp:
          type: string
          format: date-time
    ApiResponseListSalesOrderDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesOrderDto"
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListDealerLookupResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/DealerLookupResponse"
        timestamp:
          type: string
          format: date-time
    DealerLookupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        outstandingBalance:
          type: number
        creditLimit:
          type: number
    ApiResponseListCreditRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/CreditRequestDto"
        timestamp:
          type: string
          format: date-time
...